# api-gateway/api/docs/swagger.yaml
definitions:
...
  genprotos.PartyCreate:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
      opened_date:
        type: string
      slogan:
        type: string
    type: object
  genprotos.PartyDelete:
    properties:
      id:
        type: string
    type: object
  genprotos.PartyGetAllReq:
    properties:
      name:
        type: string
      opened_date:
        type: string
    type: object
  genprotos.PartyGetById:
    properties:
      id:
        type: string
    type: object
  genprotos.PartyUpdate:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
      opened_date:
        type: string
      slogan:
        type: string
    type: object
  genprotos.PublicCreate:
    properties:
      birthday:
        type: string
      email:
        type: string
      gender:
        type: string
      id:
        type: string
      last_name:
        type: string
      name:
        type: string
      party_id:
        type: string
      phone:
        type: string
    type: object
  genprotos.PublicDelete:
    properties:
      id:
        type: string
    type: object
  genprotos.PublicGetAllReq:
    properties:
      birthday:
        type: string
      email:
        type: string
      gender:
        type: string
      last_name:
        type: string
      name:
        type: string
      party_id:
        type: string
      phone:
        type: string
    type: object
  genprotos.PublicGetById:
    properties:
      id:
        type: string
    type: object
  genprotos.PublicUpdate:
    properties:
      birthday:
        type: string
      email:
        type: string
      gender:
        type: string
      id:
        type: string
      last_name:
        type: string
      name:
        type: string
      party_id:
        type: string
      phone:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /parties:
    get:
      consumes:
      - application/json
      description: Get all parties
      operationId: getallparties
      produces:
      - application/json
      responses:
        "200":
          description: party data
          schema:
            $ref: '#/definitions/genprotos.PartyGetAllReq'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: party not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Get all parties
      tags:
      - party
  /party:
    post:
      consumes:
      - application/json
      description: Create parties
      operationId: createparty
      parameters:
      - description: Party data
        in: body
        name: party
        required: true
        schema:
          $ref: '#/definitions/genprotos.PartyCreate'
      produces:
      - application/json
      responses:
        "200":
          description: party created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: party not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Create parties
      tags:
      - party
  /party/{id}:
    delete:
      consumes:
      - application/json
      description: Delete party
      operationId: deleteparty
      parameters:
      - description: party ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: party data
          schema:
            $ref: '#/definitions/genprotos.PartyDelete'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: party not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Delete party
      tags:
      - party
    get:
      consumes:
      - application/json
      description: Get party by ID
      operationId: getparty
      parameters:
      - description: party ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: party data
          schema:
            $ref: '#/definitions/genprotos.PartyGetById'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: party not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Get Party By ID
      tags:
      - party
    put:
      consumes:
      - application/json
      description: Update party
      operationId: updateparty
      parameters:
      - description: Party id
        in: path
        name: id
        required: true
        type: string
      - description: Party data
        in: body
        name: public
        required: true
        schema:
          $ref: '#/definitions/genprotos.PartyUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: party data
          schema:
            $ref: '#/definitions/genprotos.PartyUpdate'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: party not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Update party
      tags:
      - party
...
  /public/:
    post:
      consumes:
      - application/json
      description: create public
      operationId: createpublics
      parameters:
      - description: Public data
        in: body
        name: public
        required: true
        schema:
          $ref: '#/definitions/genprotos.PublicCreate'
      produces:
      - application/json
      responses:
        "200":
          description: Public data
          schema:
            $ref: '#/definitions/genprotos.PublicCreate'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Public not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: create public
      tags:
      - Public
  /public/{id}:
    delete:
      consumes:
      - application/json
      description: Delete public
      operationId: deletepublic
      parameters:
      - description: Public ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Public data
          schema:
            $ref: '#/definitions/genprotos.PublicDelete'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Public not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Delete public
      tags:
      - Public
    get:
      consumes:
      - application/json
      description: Get Public by ID
      operationId: getpublicbyid
      parameters:
      - description: Public ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Public data
          schema:
            $ref: '#/definitions/genprotos.PublicGetById'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Public not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Get Public By ID
      tags:
      - Public
    put:
      consumes:
      - application/json
      description: Update public
      operationId: updatepublic
      parameters:
      - description: Public ID
        in: path
        name: id
        required: true
        type: string
      - description: Public data
        in: body
        name: public
        required: true
        schema:
          $ref: '#/definitions/genprotos.PublicUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: Public data
          schema:
            $ref: '#/definitions/genprotos.PublicUpdate'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Public not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Update public
      tags:
      - Public
  /publics:
    get:
      consumes:
      - application/json
      description: Get all publics
      operationId: getallpublics
      produces:
      - application/json
      responses:
        "200":
          description: Public data
          schema:
            $ref: '#/definitions/genprotos.PublicGetAllReq'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Public not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      summary: Get all publics
      tags:
      - Public
swagger: "2.0"